OC 606 - Vault Delete for tool deletion

See merge request opsera-repo/web-portal/reactjs-client-portal!610

+++ src/components/common/actions/buttons/tool/ActionBarDeleteToolButton.jsx
@@ -36,6 +36,12 @@ function ActionBarDeleteToolButton({ toolDataObject }) {
 
   const deleteObject = async () => {
     try {
+      let vaultDeleteResponse = await toolsActions.deleteVaultRecordsForToolId(toolDataObject, getAccessToken)
+      if (vaultDeleteResponse.status !== 200) {
+        const errorMsg = `Error reported by services while deleting tool information from Vault. Please try again`;
+        toastContext.showErrorDialog(errorMsg);
+        return
+      }
       let result = await toolsActions.deleteTool(toolDataObject, getAccessToken);
       toastContext.showDeleteSuccessResultDialog("Tool");
       setShowDeleteModal(false);
